@@ -1,6 +1,6 @@
 {
   "name": "docbox",
-  "version": "1.0.5",
+  "version": "1.0.8",
   "description": "an api documentation website",
   "main": "index.js",
   "scripts": {
@@ -26,29 +26,30 @@
   "license": "ISC",
   "dependencies": {
     "babel-cli": "^6.4.0",
-    "babel-eslint": "^6.0.4",
+    "babel-eslint": "^8.0.0",
+    "babel-plugin-transform-class-properties": "^6.24.1",
     "babel-polyfill": "^6.3.14",
     "babel-preset-es2015": "^6.3.13",
     "babel-preset-react": "^6.3.13",
     "babel-preset-stage-0": "^6.3.13",
-    "babelify": "^7.2.0",
+    "babelify": "^8.0.0",
     "brfs": "^1.4.2",
-    "browserify": "^13.0.0",
+    "browserify": "^14.3.0",
     "cssnano": "^3.4.0",
-    "es6-promise": "^3.0.2",
-    "eslint": "^3.1.1",
-    "eslint-plugin-babel": "^3.0.0",
-    "eslint-plugin-react": "^5.1.1",
+    "es6-promise": "^4.1.0",
+    "eslint": "^4.1.0",
+    "eslint-plugin-babel": "^4.1.1",
+    "eslint-plugin-react": "^7.0.0",
     "github-slugger": "^1.0.1",
     "highlight.js": "^9.2.0",
     "isomorphic-fetch": "^2.2.0",
     "lodash.debounce": "^4.0.3",
     "minifyify": "^7.1.0",
-    "react": "^15.0.2",
-    "react-dom": "^15.0.2",
-    "react-pure-render": "^1.0.2",
-    "remark": "^7.0.0",
-    "remark-html": "^6.0.0",
+    "prop-types": "^15.5.9",
+    "react": "^16.1.1",
+    "react-dom": "^16.1.1",
+    "remark": "^8.0.0",
+    "remark-html": "^7.0.0",
     "remark-slug": "^4.1.0",
     "standard-changelog": "^1.0.1",
     "standard-version": "^4.0.0",
@@ -57,12 +58,12 @@
   },
   "devDependencies": {
     "babel-register": "^6.3.13",
-    "budo": "^8.2.2",
-    "cz-conventional-changelog": "1.2.0",
-    "expect": "^1.13.4",
-    "mocha": "^2.3.4",
+    "budo": "^10.0.0",
+    "cz-conventional-changelog": "2.1.0",
+    "expect": "^21.0.2",
+    "mocha": "^4.0.1",
     "to-vfile": "^2.0.0",
-    "uglifyjs": "^2.4.10"
+    "uglify-js": "^3.0.27"
   },
   "config": {
     "commitizen": {
